chore(front): switch eslint config to cjs

+++ database/migrations/2025_04_07_125625_add_price_column.php
@@ -11,9 +11,11 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::table('monitors_schools', function (Blueprint $table) {
-            $table->decimal('block_price', 10, 2)->nullable()->after('accepted_at');
-        });
+        if (Schema::hasTable('monitors_schools')) {
+            Schema::table('monitors_schools', function (Blueprint $table) {
+                $table->decimal('block_price', 10, 2)->nullable()->after('accepted_at');
+            });
+        }
     }
 
     /**
@@ -21,8 +23,10 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::table('monitors_schools', function (Blueprint $table) {
-            $table->dropColumn('block_price');
-        });
+        if (Schema::hasTable('monitors_schools')) {
+            Schema::table('monitors_schools', function (Blueprint $table) {
+                $table->dropColumn('block_price');
+            });
+        }
     }
 };
